chore(plugins/plugin-view-utilization): clean up view-utilization plugin code
Fixes #252
