+++ plugins/plugin-view-utilization/src/controller/get-pod-data.ts
@@ -16,7 +16,7 @@
 
 import { Arguments, Table, Row } from '@kui-shell/core'
 
-import parseAsSize from './parse-as-size'
+import parseAsSize from '../lib/parse'
 
 /**
  *
